refactor: group multipart lifecycle files
