+++ apps/backend/scripts/postgresIntegrations/boundaries.mjs
@@ -99,15 +99,15 @@ export const boundaryDefinitions = Object.freeze([
     migrationFileName: "0101_multipart_foreground_completion_fencing.sql",
     expectedMigrationCount: 103,
     testFiles: Object.freeze([
-      "src/mediaAssets/multipart/foregroundFencing.postgres.integration.ts",
+      "src/mediaAssets/multipart/writerLifecycle/foregroundFencing.postgres.integration.ts",
     ]),
   }),
   Object.freeze({
     migrationFileName: "0100_multipart_replacement_creation_claim.sql",
     expectedMigrationCount: 102,
     testFiles: Object.freeze([
-      "src/mediaAssets/multipart/replacementCreationClaim.postgres.integration.ts",
-      "src/mediaAssets/multipart/uploadSessionCreation.postgres.integration.ts",
+      "src/mediaAssets/multipart/creation/replacementCreationClaim.postgres.integration.ts",
+      "src/mediaAssets/multipart/creation/uploadSessionCreation.postgres.integration.ts",
     ]),
   }),
   Object.freeze({
@@ -117,8 +117,8 @@ export const boundaryDefinitions = Object.freeze([
       "src/database/deadline.postgres.integration.ts",
       "src/mediaAssets/blobLifecycle/lifecycle.postgres.integration.ts",
       "src/mediaAssets/ingestion/directIngestionApply.postgres.integration.ts",
-      "src/mediaAssets/multipart/completionReconciliation.postgres.integration.ts",
-      "src/mediaAssets/multipart/writerAbortReplay.postgres.integration.ts",
+      "src/mediaAssets/multipart/completion/completionReconciliation.postgres.integration.ts",
+      "src/mediaAssets/multipart/writerLifecycle/writerAbortReplay.postgres.integration.ts",
     ]),
   }),
   Object.freeze({
@@ -128,21 +128,21 @@ export const boundaryDefinitions = Object.freeze([
       "src/database/deadline.postgres.integration.ts",
       "src/mediaAssets/blobLifecycle/lifecycle.postgres.integration.ts",
       "src/mediaAssets/ingestion/directIngestionApply.postgres.integration.ts",
-      "src/mediaAssets/multipart/writerAbortReplay.postgres.integration.ts",
+      "src/mediaAssets/multipart/writerLifecycle/writerAbortReplay.postgres.integration.ts",
     ]),
   }),
   Object.freeze({
     migrationFileName: "0097_direct_multipart_writer_attempt_fencing.sql",
     expectedMigrationCount: 99,
     testFiles: Object.freeze([
-      "src/mediaAssets/multipart/writerAbortReplay.postgres.integration.ts",
+      "src/mediaAssets/multipart/writerLifecycle/writerAbortReplay.postgres.integration.ts",
     ]),
   }),
   Object.freeze({
     migrationFileName: "0096_atomic_multipart_completion_resolution.sql",
     expectedMigrationCount: 98,
     testFiles: Object.freeze([
-      "src/mediaAssets/multipart/writerAttempts.postgres.integration.ts",
+      "src/mediaAssets/multipart/writerLifecycle/writerAttempts.postgres.integration.ts",
     ]),
   }),
 ]);
