+++ apps/backend/src/entrypoints/scheduledJobs/lambda-multipart-completion-reconciliation.ts
@@ -6,7 +6,7 @@ import {
   runMultipartCompletionReconciliationBatch,
   type MultipartCompletionFailureReportBatchResult,
   type MultipartCompletionReconciliationBatchResult,
-} from "../../mediaAssets/multipart/completionReconciliation";
+} from "../../mediaAssets/multipart/completion/completionReconciliation";
 import { createCloudWatchRecord } from "../../observability/cloudWatch";
 import {
   addBackendSentryBreadcrumb,
